// src/utils/open-im-sdk-wasm/types/entity.d.ts
export declare type ConversationItem = {
  conversationID: string
  conversationType: SessionType
  userID: string
  groupID: string
  showName: string
  faceURL: string
  recvMsgOpt: MessageReceiveOptType
  unreadCount: number
  groupAtType: GroupAtType
  latestMsg: string
  latestMsgSendTime: number
  draftText: string
  draftTextTime: number
  isPinned: boolean
  isNotInGroup: boolean
  isPrivateChat: boolean
  attachedInfo: string
  ex: string
}